refactor: RepositoryImplをRepositoryにリネーム
- ArticleRepositoryImpl.ts → ArticleRepository.ts
- BookmarkRepositoryImpl.ts → BookmarkRepository.ts
- BookmarkRepositoryImpl.test.ts → BookmarkRepository.test.ts
- 全てのimport文を更新
- README.mdのディレクトリ構造図とコメントを更新
- 抽象レイヤーを除去してシンプルな実装に変更

Fixes #111
