+++ graphql/src/application/articles/FetchArticlesUseCase.ts
@@ -1,5 +1,5 @@
 import type { Article } from "../../infrastructure/domain/Article";
-import * as articleRepository from "../../infrastructure/external/ArticleRepositoryImpl";
+import * as articleRepository from "../../infrastructure/external/ArticleRepository";
 
 export const fetchArticlesUseCase = async (
   page: number,
