+++ graphql/src/application/bookmarks/DeleteBookmarkUseCase.ts
@@ -1,4 +1,4 @@
-import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepository";
 
 export const deleteBookmarkUseCase = async (id: string): Promise<boolean> => {
   return bookmarkRepository.deleteBookmark(id);
