+++ graphql/src/application/bookmarks/FetchBookmarksUseCase.ts
@@ -1,5 +1,5 @@
 import type { Bookmark } from "../../infrastructure/domain/Bookmark";
-import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepository";
 
 export const fetchBookmarksUseCase = async (): Promise<Bookmark[]> => {
   return bookmarkRepository.findMany();
