+++ graphql/src/application/bookmarks/UpdateBookmarkUseCase.ts
@@ -2,7 +2,7 @@ import type {
   Bookmark,
   UpdateBookmarkInput,
 } from "../../infrastructure/domain/Bookmark";
-import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../infrastructure/persistence/BookmarkRepository";
 
 export const updateBookmarkUseCase = async (
   id: string,
